+++ src/test/10.localBranch.test.ts
@@ -1,23 +1,22 @@
 import {
-	CustomTreeSection,
-	EditorView,
-	InputBox,
-	VSBrowser,
-	WebDriver,
-	Workbench,
+    CustomTreeSection,
+    EditorView,
+    InputBox,
+    VSBrowser,
+    WebDriver,
+    Workbench,
 } from "vscode-extension-tester";
 import { messages } from "../utils/common/messages";
-
 import {expect} from "chai";
-import {getQuickPickSelector, initialize} from "./utils/utils";
+import {getQuickPickSelector, initialize, sleep, waitForNotificationWithTimeout} from "./utils/utils";
 import {
-	BRANCH_KEY_TREE,
-	CX_CLEAR,
-	CX_SELECT_BRANCH,
-	CX_SELECT_PROJECT,
-	CX_SELECT_SCAN,
-	PROJECT_KEY_TREE,
-	SCAN_KEY_TREE_LABEL,
+    BRANCH_KEY_TREE,
+    CX_CLEAR,
+    CX_SELECT_BRANCH,
+    CX_SELECT_PROJECT,
+    CX_SELECT_SCAN,
+    PROJECT_KEY_TREE,
+    SCAN_KEY_TREE_LABEL,
 } from "./utils/constants";
 import { CX_TEST_SCAN_PROJECT_NAME } from "./utils/envs";
 import { constants } from "../utils/common/constants";
@@ -26,21 +25,21 @@ import * as fs from "fs";
 
 
 function retryTest(testFn, retries = 3) {
-	return async function () {
-		let lastError;
-		for (let attempt = 1; attempt <= retries; attempt++) {
-			try {
-				await testFn();
-				return;
-			} catch (error) {
-				lastError = error;
-				console.warn(`Retrying test... Attempt ${attempt} of ${retries}`);
-				if (attempt === retries) {
-					throw lastError;
-				}
-			}
-		}
-	};
+    return async function () {
+        let lastError;
+        for (let attempt = 1; attempt <= retries; attempt++) {
+            try {
+                await testFn();
+                return;
+            } catch (error) {
+                lastError = error;
+                console.warn(`Retrying test... Attempt ${attempt} of ${retries}`);
+                if (attempt === retries) {
+                    throw lastError;
+                }
+            }
+        }
+    };
 }
 
 function switchToBranch(branchName: string) {
@@ -54,17 +53,13 @@ function switchToBranch(branchName: string) {
 
 }
 
-async function sleep(ms: number) {
-	return await new Promise(resolve => setTimeout(resolve, ms));
-}
-
 async function selectItem(text) {
-	const input = await InputBox.create();
-	await input.setText(text);
-	let item = await getQuickPickSelector(input);
-	await input.setText(item);
-	await input.confirm();
-	return item;
+    const input = await InputBox.create();
+    await input.setText(text);
+    let item = await getQuickPickSelector(input);
+    await input.setText(item);
+    await input.confirm();
+    return item;
 }
 
 async function waitForNotificationWithTimeout(timeout) {
@@ -90,14 +85,14 @@ async function waitForNotificationWithTimeout(timeout) {
 }
 
 describe("Using a local branch if Git exists", () => {
-	let bench: Workbench;
-	let treeScans: CustomTreeSection;
-	let driver: WebDriver;
-	let originalBranch: string | undefined;
-	let gitExistedBefore: boolean;
+    let bench: Workbench;
+    let treeScans: CustomTreeSection;
+    let driver: WebDriver;
+    let originalBranch: string | undefined;
+    let gitExistedBefore: boolean;
 
-	before(async function () {
-		this.timeout(300000);
+    before(async function () {
+        this.timeout(300000);
 
 		const branchName = "branch-not-exist-in-cx";
 		// check if git repository exists
@@ -111,10 +106,10 @@ describe("Using a local branch if Git exists", () => {
 		}
 		switchToBranch(branchName);
 
-		bench = new Workbench();
-		driver = VSBrowser.instance.driver;
-		await bench.executeCommand(CX_SELECT_SCAN);
-	});
+        bench = new Workbench();
+        driver = VSBrowser.instance.driver;
+        await bench.executeCommand(CX_SELECT_SCAN);
+    });
 
 
 	after(async function () {
